requirements.txt
Flask==2.3.2
pandas==2.1.2
scikit-learn==1.3.2
joblib==1.3.2
python-docx==0.8.11
pypdf==4.2.0
Werkzeug==2.3.6
